- Refine the schemas of `Card` and `Mechanic`
- Add more unit tests for `Card` and `Mechanic`

+++ core/src/main/java/com/hearthintellect/dao/CardRepository.java
@@ -7,7 +7,7 @@ import java.util.Iterator;
 
 public interface CardRepository {
 
-    Card findById(long cardId);
+    Card findById(int cardId);
 
     Iterator<Card> findAllByName(String name);
